Nexus/requirements.txt
pywinauto>=0.6.8
pillow>=9.5.0
pytesseract>=0.3.10
